Correct the non-logging of the Integer setpoints and parameters

+++ src/main/kotlin/ch/hevs/cloudio/cloud/services/InfluxUpdateSetService.kt
@@ -46,7 +46,7 @@ class InfluxUpdateSetService(
                 // complete the point depending on attribute type
                 when (attribute.type) {
                     AttributeType.Boolean -> point.addField("value", attribute.value as Boolean)
-                    AttributeType.Integer -> point.addField("value", attribute.value as Int)
+                    AttributeType.Integer -> point.addField("value", attribute.value as Long)
                     AttributeType.Number -> {
                         if (attribute.value is Int) {
                             attribute.value = (attribute.value as Int).toFloat()
